Add back in logic for remote host and prioritize most specific
The logic for use of remote hosts was reverted when #85 was reverted.
This is needed for checks when holland is pointed to a remote host or
dbaas instance.

Also, we should be prioritizing the most specific configurations. The
credentials specified in the backupset configuration files should take
precedence as this allows one holland instance to point to multiple
hosts for backupsets.