Pull backupsets from config file rather than fixing default
This re-factors the logic to read the backupsets from the configuration
file rather than just saying "default" as the backupset.  This also
allows for monitoring of a Holland instance that has multiple backupsets
in the configuration file.  The logic currently only provides the worst
case of the configured backupsets as the value for the alert logic.
These values could come from different backupsets if multiple are
configured.